+++ Test/rankHTML.test.js
@@ -1,5 +1,5 @@
 // Test/rankHTML.test.js
-import { testHTMLFile } from './sharedHTML.js'; // Corrected import path
+import { testHTMLFile } from '../sharedHTML'; // Corrected import path
 
 describe('RankHTML Tests', () => {
   testHTMLFile('../ranksings.html', 'Ranking System'); // Adjusted path to HTML file
